pyproject.toml
[project]
name = "encode-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "portia-sdk-python[tools-browser-browserbase]",
]
